Create ValidatingPhoneNumber.py
A valid mobile number is a ten-digit number starting with 7, 8, or 9.
The first line contains an integer N, the number of inputs.
Constraints:
1<=N<=10
2<=len(Number)<=15
For every string listed, print "YES" if it is a valid mobile number and "NO" if it is not on separate lines. Do not print the quotes.